feature(entities): move to domain package and add auditable fields
