+++ src/main/kotlin/com/jlima/bookstoremanager/providers/entity/domain/AuthorEntity.kt
@@ -1,5 +1,7 @@
-package com.jlima.bookstoremanager.providers.entity
+package com.jlima.bookstoremanager.providers.entity.domain
 
+import com.jlima.bookstoremanager.providers.entity.AuditableEntity
+import org.hibernate.annotations.CreationTimestamp
 import java.util.Date
 import java.util.UUID
 import javax.persistence.Column
@@ -29,5 +31,18 @@ data class AuthorEntity(
     val birthDate: Date,
 
     @OneToMany(mappedBy = "author", fetch = FetchType.LAZY)
-    val books: List<BookEntity>
-)
+    val books: List<BookEntity>,
+
+    @Column(name = "created_at", nullable = false, updatable = false)
+    @Temporal(TemporalType.TIMESTAMP)
+    @CreationTimestamp
+    override val createdAt: Date? = null,
+
+    @Column(name = "updated_at", nullable = true, updatable = true)
+    @Temporal(TemporalType.TIMESTAMP)
+    override val updatedAt: Date? = null,
+
+    @Column(name = "deleted_at", nullable = true, updatable = true)
+    @Temporal(TemporalType.TIMESTAMP)
+    override val deletedAt: Date? = null
+) : AuditableEntity()
