+++ src/main/kotlin/com/jlima/bookstoremanager/providers/entity/domain/BookEntity.kt
@@ -1,5 +1,8 @@
-package com.jlima.bookstoremanager.providers.entity
+package com.jlima.bookstoremanager.providers.entity.domain
 
+import com.jlima.bookstoremanager.providers.entity.AuditableEntity
+import org.hibernate.annotations.CreationTimestamp
+import org.hibernate.annotations.UpdateTimestamp
 import java.util.Date
 import java.util.UUID
 import javax.persistence.CascadeType
@@ -9,6 +12,8 @@ import javax.persistence.GeneratedValue
 import javax.persistence.Id
 import javax.persistence.ManyToOne
 import javax.persistence.Table
+import javax.persistence.Temporal
+import javax.persistence.TemporalType
 
 @Entity
 @Table(name = "book", schema = "domain")
@@ -40,5 +45,20 @@ data class BookEntity(
     val publisher: PublisherEntity,
 
     @ManyToOne(cascade = [CascadeType.MERGE])
-    val user: UserEntity
-)
+    val user: UserEntity,
+
+    @Column(name = "created_at", nullable = false, updatable = false)
+    @Temporal(TemporalType.TIMESTAMP)
+    @CreationTimestamp
+    override val createdAt: Date,
+
+    @Column(name = "updated_at", nullable = true, updatable = true)
+    @Temporal(TemporalType.TIMESTAMP)
+    @UpdateTimestamp
+    override val updatedAt: Date? = null,
+
+    @Column(name = "deleted_at", nullable = true, updatable = true)
+    @Temporal(TemporalType.TIMESTAMP)
+    @UpdateTimestamp
+    override val deletedAt: Date? = null
+) : AuditableEntity()
